lyzrboost/requirements_demo.txt
streamlit>=1.22.0
lyzrboost==0.1.0
requests>=2.25.1
pyyaml>=5.4.1
google-generativeai>=0.4.0 